[TASK] split up of server manager: physics engine and mexec as separate modules
towards starting physics engine separatly from tango devices
So it can be called independently from server manager
